chore: refactor AppShell content
