--- src/themes/components/ToneWheel.jsx
@@ -1,1 +1,0 @@
-//TODO: Wheel menu inside Modal Content for pitch selection
